refactor-serialization: Split up the Memory Controller Files
