+++ src/memoryctrl/mc_frontend.h
@@ -29,60 +29,16 @@
  *
  ***************************************************************************/
 
-#ifndef MEMORYCTRL_H_
-#define MEMORYCTRL_H_
+#ifndef __MC_FRONTEND_H__
+#define __MC_FRONTEND_H__
 
 #include "XML_Parse.h"
 #include "array.h"
 #include "basic_components.h"
 #include "logic.h"
-#include "mc_backend.h"
-#include "mc_phy.h"
 #include "parameter.h"
 
 #include <vector>
-
-#if 0
-class MCBackend : public Component {
-public:
-  InputParameter l_ip;
-  uca_org_t local_result;
-  enum MemoryCtrl_type mc_type;
-  MCParam mcp;
-  statsDef tdp_stats;
-  statsDef rtp_stats;
-  statsDef stats_t;
-  powerDef power_t;
-  MCBackend(InputParameter *interface_ip_,
-            const MCParam &mcp_,
-            enum MemoryCtrl_type mc_type_);
-  void compute();
-  void computeEnergy(bool is_tdp = true);
-  void displayEnergy(uint32_t indent = 0, int plevel = 100, bool is_tdp = true);
-  ~MCBackend(){};
-};
-#endif
-
-#if 0
-class MCPHY : public Component {
-public:
-  InputParameter l_ip;
-  uca_org_t local_result;
-  enum MemoryCtrl_type mc_type;
-  MCParam mcp;
-  statsDef tdp_stats;
-  statsDef rtp_stats;
-  statsDef stats_t;
-  powerDef power_t;
-  MCPHY(InputParameter *interface_ip_,
-        const MCParam &mcp_,
-        enum MemoryCtrl_type mc_type_);
-  void compute();
-  void computeEnergy(bool is_tdp = true);
-  void displayEnergy(uint32_t indent = 0, int plevel = 100, bool is_tdp = true);
-  ~MCPHY(){};
-};
-#endif
 
 class MCFrontEnd : public Component {
 public:
@@ -104,24 +60,4 @@ public:
   ~MCFrontEnd();
 };
 
-class MemoryController : public Component {
-public:
-  ParseXML *XML;
-  InputParameter interface_ip;
-  enum MemoryCtrl_type mc_type;
-  MCParam mcp;
-  MCFrontEnd *frontend;
-  MCBackend transecEngine;
-  MCPHY PHY;
-  Pipeline *pipeLogic;
-
-  // clock_network clockNetwork;
-  MemoryController(ParseXML *XML_interface,
-                   InputParameter *interface_ip_,
-                   enum MemoryCtrl_type mc_type_);
-  void set_mc_param();
-  void computeEnergy(bool is_tdp = true);
-  void displayEnergy(uint32_t indent = 0, int plevel = 100, bool enable = true);
-  ~MemoryController();
-};
-#endif /* MEMORYCTRL_H_ */
+#endif // __MC_FRONTEND_H__
